requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
httpx==0.27.2
supabase==2.7.4
anthropic==0.34.2
google-generativeai==0.8.1
celery==5.4.0
redis==5.0.8
pydantic==2.8.2
pydantic-settings==2.4.0
firecrawl-py==1.1.1
langchain==0.3.1
langchain-anthropic==0.2.1
langchain-google-genai==2.0.1
langchain-community==0.3.1
tiktoken==0.7.0
python-multipart==0.0.9
schedule==1.2.2
